requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.2.7
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
